Move URI decoding back to Connection model
In effort to eliminate duplicate code.
